self critic
When starting this project I already knew the type of webpage I wanted to create, I knew I wanted a more modern and simplistic look. I started with the homepage, and tested all types of designs on the web page. I tried out different navigation orientations, and colors, but decided to just keep it lighter gray and left it fixed on the left side,so no matter how the page changes or moves, or if you scroll you will be able to navigate. If I had more time I would have most definitely liked to add more features to the website, an example would be adding subsections to the navigation on the left, so you can easily go through a page without the need of scrolling. I would also like to add icons or little images as I feel like the webpage is a little empty.